The bare minimum.
i understand that its an assignment - but i didn't do this one for so long because its so out of left field, you're told to *MAKE A GAME*, essentially. or at least thats how i saw it. And, i can't make something like that without a real idea. some sort of guidance, which I did not have. I finally settled on making a barebones nothing, for the sake of a grade.